WIP: FileManager: Shrink FileEntryRef to the size of a pointer
This is a WIP patch to shrink FileEntryRef to the size of a pointer,
posting for early feedback in case someone has some. This pulled in more
changes than I expected, and I'll try to find some things to split off
and commit incrementally, but I think this concept is sound
(`BasicTests` passes, still have to build and test the rest of clang).

Differential Revision: https://reviews.llvm.org/D89488