ADT: Fix reference invalidation in SmallVector::emplace_back
Avoid reference invalidation during `SmallVector::emplace_back` when one
of the arguments is an internal reference to storage and the vector has
to grow.

For POD-like types, reimplement `emplace_back` in terms of `push_back`.

For other types, split the grow operation in two and construct the new
element in the middle.

`mallocForGrow()` calculates the new capacity and returns the result of
`safe_malloc()`. We only need a single definition per `SmallVectorBase`
so this is defined in SmallVector.cpp to avoid code size bloat. Moving
this part of non-POD grow to the source file also allows the logic to be
easily shared with `grow_pod`, and `report_size_overflow()` and
`report_at_maximum_capacity()` can move there too.

`finishGrowingWithAllocationAndCapacity()` finishes the grow operation;
it still needs to be in the header. I've defined it out-of-line to
discourage inlining, similar to `grow()` itself, since growing generally
is on the slow path.

`grow()` is now implemented in terms of the two new functions.

ADT: Fix reference invalidation in size+value version of SmallVector::assign

Reimplement `SmallVector:assign(size_type, const T&)` to avoid
invalidating the argument during `clear()` when it's an internal
reference.

The new semantics match what libc++ does for `std::vector::assign`.

There is a minor semantic change when not growing. Previously, assign
would start with `clear()`, and so the old elements were destructed and
all elements of the new vector were copy-constructed.  The new
implementation skips destruction and uses copy-assignment for the prefix
of the new vector less than `size()`.

In order to avoid adding unnecessary copies, this also splits up the
`grow()` operation.

Note that the following is another possible implementation:
```
  void assign(size_type NumElts, ValueParamT Elt) {
    std::fill_n(this->begin(), std::min(NumElts, this->size()), Elt);
    this->resize(NumElts, Elt);
  }
```
The downside of this is that if the vector has to grow there will be
`size()` redundant copy operations.

Differential Revision: https://reviews.llvm.org/D94739