ADT: Sink the guts of StringMapEntry::Create into StringMapEntryBase
Sink the interesting parts of StringMapEntry::Create into a new function
StringMapEntryBase::allocateWithKey that's only templated on the
allocator, taking the entry size and alignment as parameters.

A possible follow-up would be to type-erase the allocator by passing an
`llvm::function_ref<void *(size_t, size_t)>` to allocateWithKeyImpl,
whose definition could move to  StringMap.cpp, but I'm not sure that'd
be worth it.