Support: Add WorkingDirectoryState to improve Windows CWD fidelity
WorkingDirectoryState encapsulates the current working directory, as
well as working directories for each active drive. It only supports
drives when in path::Style::windows (or ::native on Windows).

WorkingDirectoryState::makeAbsolute and
WorkingDirectoryState::isAbsolute are more precise than
sys::path::make_absolute and sys::path::is_absolute, respectively,
rejecting a path on Windows where the first component is not a valid
drive but happens to end in a colon (e.g., 'prn:').